backend/requirements.txt
psycopg2-binary
python-dotenv
python-multipart
PyJWT
passlib[bcrypt]
bcrypt==3.2.0
pytest
httpx
